+++ src/style.scss
@@ -48,9 +48,10 @@
   position: absolute;
   top: 30%;
   left: 50%;
-  border: 16px solid #0f0;     //#f3f3f3;
+  border: 16px solid #f3f3f3;
   border-radius: 50%;
   border-top: 16px solid rgba(41, 34, 34, 0.9);
+  border-bottom: 16px solid rgba(41, 34, 34, 0.9);
   width: 120px;
   height: 120px;
   -webkit-animation: spin 2s linear infinite; /* Safari */
